+++ packages/http-server/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.1](https://github.com/ttoss/ttoss/compare/@ttoss/http-server@0.3.0...@ttoss/http-server@0.3.1) (2025-12-28)
+
+### Bug Fixes
+
+- import and require ([4404d5d](https://github.com/ttoss/ttoss/commit/4404d5dd48645e417ca89ff2b959b0f2b526d33b))
+
+# 0.3.0 (2025-12-27)
+
+### Features
+
+- create http-server mcp ([#819](https://github.com/ttoss/ttoss/issues/819)) ([1566679](https://github.com/ttoss/ttoss/commit/156667902440d0de2ae2fc4729b2d5c72ddc8604))
+
+# 0.2.0 (2025-12-27)
+
+### Features
+
+- create http-server mcp ([#819](https://github.com/ttoss/ttoss/issues/819)) ([1566679](https://github.com/ttoss/ttoss/commit/156667902440d0de2ae2fc4729b2d5c72ddc8604))
+
 ## [0.1.16](https://github.com/ttoss/ttoss/compare/@ttoss/http-server@0.1.15...@ttoss/http-server@0.1.16) (2025-11-22)
 
 **Note:** Version bump only for package @ttoss/http-server